Reorganize documentation to include namespace
One of the quirks of Doxygen is that you have to have a file doc comment on a
global function, unless you have namespace documentation. I added mesh.h to
document the mesh namespace, so that I wouldn't have to have file documentation,
as the file documentation was fairly redundant with the function documentation.