Make object library includes transitive
See https://stackoverflow.com/q/38832528
CMake object library includes are not transitive for the purpose of exporting. I had to implement a workaround where I point the static/shared library at the object library's includes.

It may make more sense to eventually split this project into two: one for the C++ library and the other for the Python bindings. That way, I wouldn't have to deal with object libraries, which seem to be generally discouraged unless absolutely necessary.